refactor: move documents functions to a specific folder
